+++ Screens/Signup/SignupScreen.js
@@ -1,8 +1,7 @@
 import React, { useState } from "react";
-import { auth, db } from "../firebase_config";
+import { auth, db } from "../../firebase_config";
 import { createUserWithEmailAndPassword } from "firebase/auth";
 import { collection, addDoc } from "firebase/firestore";
-import styles from "./SignupStyle";
 import {
   View,
   Text,
@@ -10,9 +9,10 @@ import {
   Pressable,
   Image,
   KeyboardAvoidingView,
+  StyleSheet,
 } from "react-native";
 
-function SignupComponent({ navigation }) {
+function SignupScreen({ navigation }) {
   const usersRef = collection(db, "user");
 
   const [email, setEmail] = useState("");
@@ -113,4 +113,44 @@ function SignupComponent({ navigation }) {
   );
 }
 
-export default SignupComponent;
+export default SignupScreen;
+
+const styles = StyleSheet.create({
+  signupContainer: {
+    flexDirection: "row",
+    justifyContent: "center",
+    marginTop: 30,
+  },
+  logo: {
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 5,
+  },
+  container: {
+    backgroundColor: "white",
+    paddingTop: 50,
+    paddingHorizontal: 12,
+  },
+  wrapper: {
+    marginTop: 80,
+  },
+  input: {
+    borderRadius: 4,
+    padding: 12,
+    backgroundColor: "#FAFAFA",
+    marginBottom: 10,
+    borderWidth: 1,
+  },
+  button: {
+    backgroundColor: "#0096F6",
+    alignItems: "center",
+    justifyContent: "center",
+    minHeight: 42,
+    borderRadius: 4,
+  },
+  buttonText: {
+    fontWeight: "600",
+    color: "#fff",
+    fontSize: 20,
+  },
+});
