--- Signup/SignupStyle.js
@@ -1,43 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  signupContainer: {
-    flexDirection: "row",
-    justifyContent: "center",
-    marginTop: 30,
-  },
-  logo: {
-    alignItems: "center",
-    justifyContent: "center",
-    borderRadius: 5,
-  },
-  container: {
-    backgroundColor: "white",
-    paddingTop: 50,
-    paddingHorizontal: 12,
-  },
-  wrapper: {
-    marginTop: 80,
-  },
-  input: {
-    borderRadius: 4,
-    padding: 12,
-    backgroundColor: "#FAFAFA",
-    marginBottom: 10,
-    borderWidth: 1,
-  },
-  button: {
-    backgroundColor: "#0096F6",
-    alignItems: "center",
-    justifyContent: "center",
-    minHeight: 42,
-    borderRadius: 4,
-  },
-  buttonText: {
-    fontWeight: "600",
-    color: "#fff",
-    fontSize: 20,
-  },
-});
-
-export default styles
