+++ package.json
@@ -12,6 +12,8 @@
     "expo-application": "~4.0.1",
     "expo-camera": "~12.1.0",
     "expo-cli": "^5.0.3",
+    "expo-clipboard": "~2.1.0",
+    "expo-image-picker": "~12.0.1",
     "firebase": "^9.6.5",
     "firebaseui": "^0.600.0",
     "react": "17.0.1",
@@ -28,9 +30,7 @@
     "react-native-web": "0.17.1",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "^2.10.4",
-    "reanimated-bottom-sheet": "^1.0.0-alpha.22",
-    "expo-image-picker": "~12.0.1",
-    "expo-clipboard": "~2.1.0"
+    "reanimated-bottom-sheet": "^1.0.0-alpha.22"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
